Renamed Search tab components
